requirements.txt
Django
djangorestframework
pytz
django-tables2
gunicorn